Fix warning

+++ vt/src/vt/GLTileRenderer.cpp
@@ -1652,7 +1652,7 @@ namespace carto { namespace vt {
             glUniform1f(glGetUniformLocation(shaderProgram, "uGamma"), gamma);
         }
         else if (geometry->getType() == TileGeometry::Type::POLYGON3D) {
-            double tileHeightScale = cglib::length(cglib::transform_vector(cglib::vec3<double>(0, 0, 1), calculateTileMatrix(tileId)));
+            float tileHeightScale = static_cast<float>(cglib::length(cglib::transform_vector(cglib::vec3<double>(0, 0, 1), calculateTileMatrix(tileId))));
             glUniform1f(glGetUniformLocation(shaderProgram, "uUVScale"), 1.0f / vertexGeomLayoutParams.texCoordScale);
             glUniform1f(glGetUniformLocation(shaderProgram, "uHeightScale"), blend / vertexGeomLayoutParams.heightScale * vertexGeomLayoutParams.coordScale);
             glUniform1f(glGetUniformLocation(shaderProgram, "uAbsHeightScale"), blend / vertexGeomLayoutParams.heightScale * POLYGON3D_HEIGHT_SCALE * tileHeightScale);
